cloud_functions/requirements.txt
functions-framework==3.*
pymongo==3.9.0
dnspython
requests
cryptography
pyjwt